refactor(#78): 로비 조회 유스케이스 서비스 분리
